Re-enable benchmarks for minDepVersions
Benchmarks are able to run again but only for minDepVersions
builds. Due to the way benchmarks are setup in the support
library android plugin (rightfully labled as "disgusting"
in the comments"), enabling two installations of benchmark
apks (minDepVersions and maxDepVersions) is not intuitive.

Since we ship the minDepVersions flavor, benchmarks are enabled
for that flavor for now. DO NOT RUN connectedCheck and instead run
connectedMinDepVersionsDebugAndroidTest.

Test: ./gradlew slice-benchmark:connectedMinDepVersionsDebugAndroidTest

Bug: 119300891
Change-Id: I1617ba4d82fad9a3367b8bbdc0996c5689cd51cb