Remove OnNavigatorNavigatedListener entirely
Instead of requiring each Navigator to send
dispatch events, have Navigator.navigate()
return which destination was navigated to
(or null if the navigate failed).

This simplifies the logic of all Navigators,
removes a common source of confusion in
implementing a Navigator, and gives
NavController a more direct way to update its
back stack.

With this change, NavController now always
ensures that all parent NavGraphs are on
the back stack after a navigate() event,
avoiding IllegalArgumentExceptions.

Test: lots of updated tests
BUG: 111278637
Change-Id: I66f6ee3ab37b56961388bd1f607b14a713f830fe