Support for custom rules
Inspired from (aosp/723326), this adds support for custom lint
checks that makes it a matter of just adding a class and then
putting it in the AndroidXIssueRegistry list.

Two errorprone rules are now also enforced in lint:
BanKeepAnnotation and BanParcelableUsage. Errorprone may soon
be disabled if no need for it remains.

Current violations are suppressed.

Test: ./gradlew lintMinDVD
Change-Id: I9998c86a0b69a23aa6aa7a5590d935951a27f9dd