[WebView Support Library] fix API 1.0.0 txt file
This removes 2 APIs which were incorrectly added during the switch to
metalava (http://aosp/814037).

That CL was based on the state of src directories at
1529fd044d504ba2bf46f22739a6670fa3fa6192 (http://ag/4400052). However,
this is not a correct algorithm, since that commit is on the master
branch, and 1.0.0 APIs were finalized on the pi-androidx-dev branch. The
correct commit should've been b39f83e0a5f1e2a49db89c42a0b2cddc92bdc7fc
(http://ag/4369068).

This manually edits the file (verified by diffing before
http://aosp/814037), since I'm not sure how to replicate the process of
the original metalava CL with the correct commit.

Bug: 120172279
Test: N/A
Change-Id: I92728e05a8bfb8481c12522f9a00a24ae640c684