Use the menuCategory to determine pop behavior
When using setupWithNavController() with a
BottomNavigationView or NavigationView, the
default behavior should be to pop the back
stack to the start destination of the graph,
but in certain cases, such as a Settings or
Help screen, you do not want this behavior.

Use menuCategory="secondary" to avoid popping
the back stack.

Test: Manual testing in the testapp
BUG: 120104424
Change-Id: Ic2567770896470726b4f12c5b4fb37ff544b6b45