Port of Inks (Ripples)
Ripples(or Inks) are drawn on top of the Material widget(added in the previous CL). It will require some adaptation for a new Crane layouts world once we will support touches there, but it is easier to start with a port and then adapt when we ready to do so. Also we will not use exactly Flutter Animation framework for ripples, but we can't just wait for it to be ready, so we can reuse Flutter's one as then we just need to port 1 to 1 and at some point later we will migrate to the new Animations framework when ready.

There are two Ink implementations, both are used in Flutter:
1) InkRipple - which is an ink splash feature that expands more aggressively than InkSplash.
2) InkSplash - which is an ink splash feature that expands less aggressively.

Bug: 114021677
Test: Added new RippleActivity to play with it. Ripple works nicely there!
Change-Id: I1c8475deb59f3da31b7d09840518b5294f652c8e