[WebView Support Library] Add test for onReceivedHttpError
This adds a test for onReceivedHttpError. This test was originally
omitted because we didn't have a web server available for AndroidX
tests. This blocker was fixed by laisminchillo@ in http://aosp/795653.

This adds okhttp3's MockWebServer library to AndroidX dependencies
(copied from laisminchillo@'s other CL: http://aosp/795891), and
reimplements the CTS test. This also adds the cleartext attribute to our
AndroidX tests because the MockWebServer is served on localhost (which
uses HTTP).

This deviates from the CTS test by checking that the mock server
actually receives the correct HTTP request. This helps distinguish if
WebView is blocking traffic (such as it would if the cleartext manifest
tag is omitted).

Bug: 119592680
Test: ./gradlew :webkit:connectedAndroidTest
Test: ./gradlew :webkit:checkExternalLicenses (with http://aosp/847912)
Change-Id: I1a666f3b570dfc30d30d4f28cbcac2f992290a58