Rename media-widget to media2-widget
The mavenGroup should be changed to MEDIA2(see the bug below). In order
to change the group, the directory should be renamed as well because
the project name should starts with "media2" to build according to
MavenUploadHelper.kt L42 and it will be confusing if the project name
and the directory path differ.

Bug: 120955709
Test: ./gradlew buildOnServer
Change-Id: Icd73ae0c3430c4ea0f72a73f48f80534f4595a17