Make SafeArgsTask depend on ApplicationIdWriterTask
This changes fixes two things:

1. The SafeArgs task not being correctly register with the variant
compile task since it was being done in the lambda of tasks.register().
In this change the task is reverted back to being create with the
create() API instead of register() since there isn't an API in AGP to
register a Java generating task using a TaskProvider.

2. The SafeArgs task executing before the ApplicationIdWriterTask which
writes the app id consumable in AGP 3.3.0+. This changes fixes this by
explicitly depending on the task that generates the TextResource
artifact that contains the application id in newer versions of the
Android plugin.

Finally since now the app id comes from two different types based on the
Android plugin version, this change also consolidates the app id @Input
of the Safe Args task into a getter.

Bug: 119662045
Test: Manual tests along with:
      ./gradlew navigation:navigation-safe-args-gradle-plugin:test

Change-Id: Ib2820d72ba52f91d51797ed2159e60497874fe80