Fix for androidx-platform-dev build failure
Metalava only tracks new methods as compared to the super class,
 this one comes from Activity#onStateNotSaved so it did not show up.
However, Activity#onStateNotSaved got deprecated in Q and so Metalava
shows it as a new method now. Hence the breakages in androidx-platform-dev
only (which uses Q SDK) and not androidx-master-dev (which uses P SDK).

Ran './gradlew updateApi' to genereate these changes for platform
branch.

Test: ./gradlew checkApi
Bug: b/121373749
Change-Id: Ie1ecfb4869967b79290621ad28ee2974e4f84585