Media, Media2 Test: Fix failures when the alarm volume is min value
When the alarm volume is min value, the volume value could be 1. In that
case, tests would fail because the tests try to lower the volume to 0
while it's impossible.

This CL fixes the failure by setting the min volume from 0 to 1 for
STREAM_ALARM. Although the exact min volume can be varies per device
configuration (e.g. STREAM_ALARM's min volume is 0 on TV devices), but
it's sufficient for tests.

Note that AudioManager.getMinStreamVolume() cannot be used because
it's added in API 28.

Test: Set alarm stream volume to min and run following command
      $ ./media/version-compat-test/runtest.sh 4
      $ ./media2/version-compat-test/runtest.sh 1
Change-Id: Ib9afa767b97bfe885ab5249838fc515f604b399d