Migrate Inks and Material to R4a
Use R4a syntax for Material, InkResponse and InkWell widgets. To not be blocked and not force Shep with touches implementation I created simple implementation in InkTmpGestureDetector to replace with the real one once it would be ready.

Material package from ui-port migrated into ui-material package. Additional core changes:
In Layout:
* Introduced LayoutCoordinates concept. It is a holder of measured bounds(and potentially matrix)
* Introduced OnPositioned composable. It's children lambda will be called after a composition (as measured bounds only available after the composition) and will provide the result LayoutCoordinates.
In Dimension:
* Added lerp methods
* Some additional methods like Dimenstion.unaryMinus, Position.minus(), Position.getDistance, Size.center()
* Added a new entity Bounds - a Rect with Dimension units

Bug: 114021677
Test: a new Inks activity to test the component
Change-Id: Ic99b082952d0f8618f87e85dcea9cf11cdd1b45a