Add className for recyclerView.
I removed this because there were no user clients for this.
But apparantly some automation tests in google3 use this.
So add it back in, but unlike the original code, let's not have this break if there is a custom delegate, and break if proguard is involved in any way.

Test: it builds
Change-Id: I51513e13a6b97bccc096b6e304e86c53bd08f976