Make source transformer operate like jetifier
The changed file (rewritePackageNames.py) will be renamed to
sourceJetifier.py in a next CL. I'll also investigate how
to make it completely in python as opposed to launching
the shell sed command that way it works for Windows as well.

The script is no longer designed to operate in AOSP, but rather
in a generic manner with the following changes:

1- Jetifier config files are now passed in to the script as an
argument by specifying the path following the -c argument.

2- The script takes in the source file path to jetify and
the path for the output file containing the jetified source.

3- The script is now also capable of dejetifying source using
the -r option.

Also there is really no more need for the print lines, especially
as this is a step towards streamlining the stream to be released
to the public soon.

Test: ./rewritePackageNames.py -i <input-java-file> -c \
../core/src/main/resources/default.generated.config

Bug: 121376473
Change-Id: I1223229d9ba1c2a7a158d9dadfe9c259a85d0696