Fix flaky FragmentScenarioTest.fragmentWithOptionsMenu
The test was flaky due to openActionBarOverflowOrOptionsMenu().
That method works well API level 28 and below but is very flaky after 29.
As for the short-term solution I added Activity#openOptionsMenu() call which
opens menu if it is not displayed yet otherwise no-op. Unfortunately, we still
need to keep both calls because openOptionsMenu() has the same flakiness on
API level 28 an below. The long-term solution should be in Espresso.

The test was also failing consistently on Android TV device because
those devices don't have menu at all. I added the check not to run the
test on Android TV devices.

Fixes: 124071322
Test: Ran FragmentScenarioTest against API level 28, 29 on Pixel 2. And Android TV with API level 24.
Change-Id: If959480202fdf501546776fd32908046b5ac3999