Add tests where fling ends in snapped position
Adds a test for ViewPager2 where a fling (drag with velocity) ends
exactly at a page. This is added to the DragWhileSmoothScrollTest
because it can only be tested when the ViewPager2 is moved to a
non-snapped position before initiating the fling. This has the benefit
that this scenario is thoroughly tested with the different parameters.

Test: ./gradlew viewpager2:connectedCheck
Change-Id: I383065081794aa0aee45d002eedb13e18318c873