Update Transition library to beta01
Plus use stable versions of dependencies instead of project(*) to not force users to upgrade to alphas of core and fragment.

Also as agreed in b/127249018 I am adding @SuppressLint("RestrictedApi") for all the usages of the methods from core and fragment libs which have LIBRARY_GROUP instead of a new LIBRARY_GROUP_PREFIX in stable versions. Created a bug b/127286008 to remove that once we would be unblocked on this.

Bug: 127249018
Bug: 127286008
Test: Compiles, sample app still works
Change-Id: I8e6e5a6b4b1065cb094259e47a2fed1627e9e652