Fix theme issues.
Remove tabStyle as a default within Theme.Car. TabLayout requires
material themes, and it is easier to allow developers to manually
specify this style on a TabLayout themselves than require it at the top
level. A developer might not even have a TabLayout and specifying this
forces a specific parent theme to have to be used.

Also, remove the attribute query within scrollbar buttons. The buttons
already look for buttonRippleBackground programmatically, so this query
is unneeded.

Fixes: 127823254
Test: Verify in SupportCarDemos
Change-Id: Iab33cffac43a557ece94fd625ed83bebe10691bd