Added BlendModeCompat APIs to leverage the BlendMode APIs on Android Q and falling back to PorterDuff.Mode equivalents wherever possible
Added APIs to PaintCompat to utilize BlendModeCompat, a shadow
implementation of the BlendMode APIs in Android Q.

FYI this was implemented as an enum to match the BlendMode APIs
added in the framework that are also implemented as an enum.
The BlendMode APIs in the framework are implemented as an enum
in order to match the original PorterDuff APIs in the platform
which are to be replaced with BlendMode.

Additionally added support to configure a BlendModeColorFilter, falling
back to PorterDuffColorFilter if BlendMode is not available.
In the event that a BlendMode is used that is not supported by
a platform version, these APIs act as no-ops

Added ktx methods on Paint to support configuring BlendModes through
BlendModeCompat

Bug: 126726419
Test: Added Tests to verify new PaintCompat BlendMode related methods
Change-Id: Iaee85ae3ad55f1548595bb8c41fea6415f588562