Return proper PrimaryNavFragment after SingleTop followed by Pop
SingleTop causes a replace and a pop to be executed in the same
transaction. If the pop is processed first, the order of operations
causes the oldPrimaryNavFragment to be lost. Doing the operations in
reverse means the oldPrimaryNavFragment gets set last and is available
for the replace operation to use.

Test: ran fragment tests and tests in class
BUG: 124332597
Change-Id: If90e2dbda0a8bd3bf8ebe91579dfee205d096356