Introduce OnChildPositioned, fix a ripple position inside the RippleSurface
Ripples wasn't positioned properly within it's RippleSurface. We should calculate a position of the internal MeasureBox(Clicked item) inside a parent MeasureBox(RippleSurface) and apply this drawing transformation.
For this case I had to implement OnChildPositioned in addition to the current OnPositioned. As this makes a lot of sense to put children composables inside OnChildPositioned(we will receive coordinates for this children) moved onPositioned callback from the @children to just a parameter and the same for OnPositioned.

Bug: 128678469
Test: Improved Ripple Demo so this cases is also visible there. Added RippleEffectTest.
Change-Id: Ibe53b28205a22f79b92c470d5cb0f7450f6868e5