Deprecate boundsInParent in favor of boundInScreen.
If the service has boundsInScreen the boundsinparent can be computed.
Also for its whole existence, boundsInParent is the bounds within the
ViewParent's coordinates. But the ViewParent is quite often not the
AccessibilityParent (if it isn't important for accessibility), so the
service can't really do anything with that information because the
service doesn't have access to the ViewParent, only the
AccessibilityParent.
A similar change has been made on the platform.

Test: It builds.
Bug: 1084662
Change-Id: I73414191f0da18496a5ee4ead57395cbe3500b07