Remove argument Bundle from FragmentFactory#instantiate
Provide a clearer separation between what
instantiate() is supposed to do (just
construct the class instance) versus what is
the responsibility of the caller (set the
arguments on the resulting Fragment).

As arguments can change at any time, it doesn't
make as much sense to use them at instantiation
time.

Test: tests still pass
Fixes: 128836103
Change-Id: I8018f4963ace5e58144ed1596c3b1e0b46638a7e