Fix racy CallbackToFutureAdapterTest#testGcedCallback
This test finishes its loop once future is resolved (isDone) and then
checks that CancellationListener is called. But it is racy, because
CallbackToFutureAdapter intentionally resolves future first and then run
cancellation listeners, so there is no guarantee that cancellation
listeners are already called after that while loop. Now test awaits
for cancellation listener instead of expecting it too be called.

bug: 120236209
Test: CallbackToFutureAdapterTest#testGcedCallback
Change-Id: I34cfd9cf5049a874b92a22d395624c2020d63db7