Add Wrap widget
The CL adds the Wrap widget, which expects a child and implements a wrap
content behavior.  Wrap attempts to be as large as its child. When this
is not possible (the child is smaller than Wrap's min constraints), the
specified alignment will be used to position the child inside Wrap (and
Wrap will size itself to min constraints).

Bug: 115463945
Test: ran WrapTest
Change-Id: I2b18475c317a5bdef2092d5a74ad6f167e73b14d