Use Color and TextStyle instead of lambdas with MaterialColor/Typography scopes
It seems like this lambdas on theme objects will not be flexible enough so probably it is better to remove them for now.
Using the color from the MaterialColor is just one use-case from many and we should not build our api on this. Developers will probably have wider color set that is possible to provide via MaterialColors, Or they even can decide to not  use our MaterialColors class at all to keep their naming to match the design system of the brand).
The usage will now be changed from:
<Button color={ secondary }/>
<StyledText style={ subtitle2 }/>
To:
<Button color=+themeColor { secondary }/>
<StyledText style=+themeTextStyle { subtitle1 }/>

Bug: 113845528
Bug: 114021677
Test: Demo app and Rally app works as before
Change-Id: I2ef933aad4e7a8e050ccd3d2ab54878b4feb321a