Dispatch onPageSelected when adapter is set
When ViewPager2 is instantiated, it should dispatch an onPageSelected
event. In the more general case, it should dispatch an onPageSelected
event whenever the adapter is set or changes.

This can be detected by the fact that RecyclerView sends an initial
onScrolled event, while the ScrollEventAdapter is IDLE. Scroll events
can't happen when it's IDLE, so it must be coming from a newly populated
RecyclerView.

Incidentally, this removes the need to notify the ScrollEventAdapter
that mCurrentItem has been restored.

Bug: 128442660
Test: ./gradlew viewpager2:connectedCheck
Change-Id: If6a4dacfa19d40730a6e0a126853111c973022cc