Cleanup sample apps
This is the first step of a cleanup.
The MainActivity allows for subdirectories by just naming the
activities with Foo/MyActivity so removed the MainActivity in
Material and Animations and used this method instead.

Also moved all r4a packaging to androidx.ui.demos instead
Deleted an unused R4AActivity class

Next steps will be moving all these classes over out of samples
if appropriate and getting rid of ui-r4a module

Test: build, run and open all activities
Change-Id: I63b8e02e60d52906f0745f4d6fb9f873e1b9f494