Ensure retained Fragments can't desync from saved state
When the instance state of the FragmentManager is
already saved, operations can still affect the set
of retained Fragments, which still need to be in
sync when the FragmentManager is recreated in
restoreSaveState().

By avoiding updating the FragmentManagerViewModel,
we ensure that we have a consistent state when we
come back up.

Test: new FragmentLifecycleTest
Fixes: 130433793
Change-Id: I66076adf50816587ae564f46e119b150e0c30360