Create a receiver scoped callback for addCallback
Rather than requiring developers construct
a OnBackPressedCallback, Kotlin users can
pass in a lambda that has a receiver scoped
OnBackPressedCallback, allowing them to access
the other methods on OnBackPressedCallback from
within their lambda.

The created OnBackPressedCallback is returned
so that it can be passed to other components or
have its methods called from outside the lambda.

Test: new tests
BUG: 129946737
Change-Id: Id510373020121621035f68fac44194d0e2584a95